Add GitHub action to create release candidates
Although the release candidate container has helped to make builds
reproducible and limit the amount of setup needed, it is still difficult
to setup in some environments. This adds a custom GitHub action and
workflow to build/publish release candidates.

To implement this, the new custom GitHub action is called
"release-candidate" and prepares the GitHub CI environment for creating
a release candidate for ASF. This includes things like SBT settings for
publishing to ASF maven repositories, creating directories for artifacts
on dist.apache.org, importing signing keys, and creating the source
artifact. Workflows using this action only need to publish artifacts
(e.g. using sbt publishSigned) and write helper binaries to a provided
artifacts directory. The post script of the action will then sign,
checksum, and commit the artifacts.

DAFFODIL-2971