Move check-release script into a container
The check release script requires a number of tools that might not be
avilable on some systems. To make it easier to test on those, this moves
the check-release script into a container that includes all the
necessary dependencies.

Note that the script can still be run without the container if the tools
exist on the system, so having the ability to run containers is not
strictly required to run the verification script. And most tools the
script uses can be installed on most systems. The one exception is
msidiff which is only needed to check build reproducibility. This also
removes options from wget that are not necessary and aren't available on
all versions of wget, making it easier to run the script without a
container.

DAFFODIL-2972