Modify check-release script to work with SBT and VS Code releases
The dist directory for SBT and VS Code releases includes an extra
'daffodil-sbt/' or 'daffodil-vscode/' directory, compared to normal
daffodil releases. And the value provided to --cut-dir does not consider
these extra directories, so downloaded directory tree structure doesn't
actaully match the structure created by the build-release container,
causing the reproducible diff to fail.

To avoid needing to detect and specify different cut dir values based on
the release type, this moves the wget logic to a new 'download_dir'
function that calculates the value of cut-dirs so that it only keeps the
last directory.

So for example, all dist artifacts are downloaded to a
directory matching the version (e.g. 1.0.0-rc1/) and all maven artifacts
are downloaded to a directory matching the maven profile (e.g.
orgapachedaffodil-1234/). Note that the build-container does not include
the maven profile directory, so after it is downloaded we move the
contents out of that directory.

This matches the structure that the build-release container creates
allowing reproducible diffs to work.

DAFFODIL-2971