Output count of total failures
With the amount of lines output, it can be difficult to notice a small
number of failures. This keeps a count of all failures and outputs the
total at the end to make it more obvious if all tests passed or not.

The main difficulty in keeping a count is that most of the checks are
done in "find ... -exec", and the exec doesn't have a way to increment a
bash variable since it occurs in a separate process space. To get around
this, we add a new "test_files" function that accepts a list of files
(retrieved using the same find command) and a string of commands to run
on each file, using {} to reference each file similar to -exec. Checking
the result of the command evaluation happens in the same process space,
so we can check for failure and increment a variable. This new function
also handles printing pass/fail results. The all means the new function
is a bit more complex, but does make the actual tests more clear.

This also changes the recursive diff to use the new test_files function,
allowing us to compare files individually so we can list both matching
and non-matching files. This makes it more clear exactly which files are
checked for reproducibility. This also switches to using cmp since we no
longer need the recursive capabilities that diff provides.

Also add missing quotes in some places and rewrite RPM signature removal
to avoid potential issues with spaces in files names.

DAFFODIL-3039