Change check-release container to Ubuntu
Something with the Fedora check-release container can lead to issues
with gpg validation, something about lock files alreay existing and
timing out. It isn't clear exactly what the problem is, but it does not
occur on Ubuntu. So this switches the check-release container to Ubuntu.
This also makes it consistent with the build-release container which is
also Ubuntu based so they can share the same base image.

The only issue with this is Ubuntu wget seems to create index.html.tmp
files that it doesn't delete--other systems do delete these files. This
seems to be a known issue with some versions of wget, so we just delete
any tmp files if they exist. This is also useful in cases where users
directly run the script on Ubuntu systems without the container.